--- app/src/main/AndroidManifest.xml
@@ -1,16 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="ru.ifmo.android_2016.calc">
-
-    <application
-        android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
-        android:theme="@style/AppTheme">
-        <activity android:name=".CalculatorActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
-                <category android:name="android.intent.category.LAUNCHER"/>
-            </intent-filter>
-        </activity>
-    </application>
-</manifest>
